dock2dns/requirements.txt
docker-py
cffi 
pyopenssl 
ndg-httpsclient 
pyasn1 